# server/openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: thomas.schaffter@sagebionetworks.org
    name: The ROCC Team
    url: https://Sage-Bionetworks.github.io/rocc-schemas
  description: |
    The OpenAPI specification implemented by the Challenge Registries.
    # Introduction
    TBA
  license:
    name: Apache 2.0
    url: https://Sage-Bionetworks.github.io/rocc-schemas
  title: Registry of Open Community Challenge API
  version: 0.3.0
  x-logo:
    url: https://Sage-Bionetworks.github.io/rocc-schemas/logo.png
servers:
- url: '{protocol}://rocc.org/api/v1'
  variables:
    protocol:
      default: https
      enum:
      - http
      - https
...
paths:
...
  /challenges:
    delete:
      description: Delete all challenges
      operationId: delete_all_challenges
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete all challenges
      tags:
      - Challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
    get:
      description: Returns all the challenges
      operationId: list_challenges
      parameters:
      - description: Maximum number of results returned
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 10
          type: integer
        style: form
      - description: Index of the first result that must be returned
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - description: Property used to sort the results that must be returned
        explode: true
        in: query
        name: sort
        required: false
        schema:
          enum:
          - createdAt
          - updatedAt
          type: string
        style: form
      - description: Can be either `asc` or `desc`. Ignored without `sort` parameter.
        explode: true
        in: query
        name: direction
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
        style: form
      - description: A string of search terms used to filter the results
        explode: true
        in: query
        name: searchTerms
        required: false
        schema:
          type: string
        style: form
      - description: Array of tag ids used to filter the results
        explode: true
        in: query
        name: tagIds
        required: false
        schema:
          items:
            $ref: '#/components/schemas/TagId'
          type: array
        style: form
      - description: Array of challenge status used to filter the results
        explode: true
        in: query
        name: status
        required: false
        schema:
          items:
            $ref: '#/components/schemas/ChallengeStatus'
          type: array
        style: form
      - description: Array of challenge platform ids used to filter the results
        explode: true
        in: query
        name: platformIds
        required: false
        schema:
          items:
            $ref: '#/components/schemas/ChallengePlatformId'
          type: array
        style: form
      - description: Return challenges that start during the date range specified
        explode: true
        in: query
        name: startDateRange
        required: false
        schema:
          $ref: '#/components/schemas/DateRange'
        style: deepObject
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfChallenges'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: List all the challenges
      tags:
      - Challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
    post:
      description: Adds a challenge
      operationId: create_challenge
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeCreateRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeCreateResponse'
          description: Success
          links:
            GetChallengeById:
              $ref: '#/components/links/GetChallengeById'
            DeleteChallengeById:
              $ref: '#/components/links/DeleteChallengeById'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request conflicts with current state of the target resource
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Add a challenge
      tags:
      - Challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenges/{challengeId}:
    delete:
      description: Deletes the challenge specified
      operationId: delete_challenge
      parameters:
      - description: The ID of the challenge
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          $ref: '#/components/schemas/ChallengeId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete a challenge
      tags:
      - Challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
    get:
      description: Returns the challenge specified
      operationId: get_challenge
      parameters:
      - description: The ID of the challenge
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          $ref: '#/components/schemas/ChallengeId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get a challenge
      tags:
      - Challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
...
  /organizations:
    delete:
      description: Delete all organizations
      operationId: delete_all_organizations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete all organizations
      tags:
      - Organization
      x-openapi-router-controller: openapi_server.controllers.organization_controller
    get:
      description: Returns the organizations
      operationId: list_organizations
      parameters:
      - description: Maximum number of results returned
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 10
          type: integer
        style: form
      - description: Index of the first result that must be returned
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfOrganizations'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get all organizations
      tags:
      - Organization
      x-openapi-router-controller: openapi_server.controllers.organization_controller
    post:
      description: Create an organization with the specified account name
      operationId: create_organization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationCreateResponse'
          description: Success
          links:
            GetOrganizationById:
              $ref: '#/components/links/GetOrganizationById'
            DeleteOrganizationById:
              $ref: '#/components/links/DeleteOrganizationById'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request conflicts with current state of the target resource
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Create an organization
      tags:
      - Organization
      x-openapi-router-controller: openapi_server.controllers.organization_controller
  /organizations/{organizationId}:
    delete:
      description: Deletes the organization specified
      operationId: delete_organization
      parameters:
      - description: The unique identifier of the organization, either the user ID
          or account name
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          oneOf:
          - $ref: '#/components/schemas/AccountId'
          - $ref: '#/components/schemas/AccountLogin'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete an organization
      tags:
      - Organization
      x-openapi-router-controller: openapi_server.controllers.organization_controller
    get:
      description: Returns the organization specified
      operationId: get_organization
      parameters:
      - description: The unique identifier of the organization, either the user ID
          or account name
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          oneOf:
          - $ref: '#/components/schemas/AccountId'
          - $ref: '#/components/schemas/AccountLogin'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get an organization
      tags:
      - Organization
      x-openapi-router-controller: openapi_server.controllers.organization_controller
...
components:
  links:
    GetChallengeById:
      description: |
        The `id` value returned in the response can be used as the `challengeId` parameter in `GET /challenges/{challengeId}`
      operationId: getChallenge
      parameters:
        challengeId: $response.body#/challengeId
    DeleteChallengeById:
      description: |
        The `id` value returned in the response can be used as the `challengeId` parameter in `DELETE /challenges/{challengeId}`
      operationId: deleteChallenge
      parameters:
        challengeId: $response.body#/challengeId
...
    GetOrganizationById:
      description: |
        The `id` value returned in the response can be used as the `organizationId` parameter in `GET /organizations/{organizationId}`
      operationId: getOrganization
      parameters:
        organizationId: $response.body#/organizationId
    DeleteOrganizationById:
      description: |
        The `id` value returned in the response can be used as the `organizationId` parameter in `DELETE /organizations/{organizationId}`
      operationId: deleteOrganization
      parameters:
        organizationId: $response.body#/organizationId
...
  schemas:
    TagId:
      description: The unique identifier of a tag
      example: 507f1f77bcf86cd799439011
      title: TagId
      type: string
    ChallengeStatus:
      description: The status of the challenge
      enum:
      - upcoming
      - active
      - completed
      example: active
      title: ChallengeStatus
      type: string
    ChallengePlatformId:
      description: The unique identifier of a challenge platform
      example: 507f1f77bcf86cd799439011
      title: ChallengePlatformId
      type: string
    DateRange:
      description: A date range
      example:
        start: 2021-07-17
        end: 2021-07-17
      properties:
        start:
          description: Start date of the date range
          format: date
          type: string
        end:
          description: End date of the date range
          format: date
          type: string
      title: DateRange
      type: object
    ResponsePageMetadata:
      description: A page of results
      properties:
        offset:
          description: Index of the first result that must be returned
          type: integer
        limit:
          description: Maximum number of results returned
          type: integer
        paging:
          $ref: '#/components/schemas/ResponsePageMetadata_paging'
        totalResults:
          description: Total number of results in the result set
          type: integer
      required:
      - limit
      - offset
      - paging
      type: object
    ChallengeId:
      description: The unique identifier of the challenge
      example: 507f1f77bcf86cd799439011
      type: string
    ChallengeCreateResponse:
      description: The unique identifier of the challenge created
      example:
        id: 507f1f77bcf86cd799439011
      properties:
        id:
          description: The unique identifier of the challenge
          example: 507f1f77bcf86cd799439011
          type: string
      required:
      - id
      title: ChallengeCreateResponse
      type: object
    ChallengeCreateRequest:
      description: The information used to create a challenge
      example:
        name: awesome-challenge
        displayName: Awesome Challenge
        description: This challenge is awesome.
        websiteUrl: https://synapse.org/awesome-challenge
        status: active
        startDate: 2020-11-10
        endDate: 2020-12-31
        platformId: 507f1f77bcf86cd799439011
      properties:
        name:
          description: The name of the challenge
          maxLength: 60
          minLength: 3
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          type: string
        displayName:
          maxLength: 60
          minLength: 3
          type: string
        description:
          description: A short description of the challenge
          maxLength: 280
          nullable: true
          type: string
        websiteUrl:
          format: uri
          type: string
        status:
          $ref: '#/components/schemas/ChallengeStatus'
        startDate:
          format: date
          nullable: true
          type: string
        endDate:
          format: date
          nullable: true
          type: string
        platformId:
          description: The unique identifier of a challenge platform
          example: 507f1f77bcf86cd799439011
          title: ChallengePlatformId
          type: string
      required:
      - description
      - displayName
      - name
      - status
      - websiteUrl
      type: object
    AccountId:
      description: The unique identifier of an account
      example: 507f1f77bcf86cd799439011
      title: AccountId
      type: string
    Challenge:
      allOf:
      - $ref: '#/components/schemas/ChallengeCreateResponse'
      - $ref: '#/components/schemas/ChallengeCreateRequest'
      - $ref: '#/components/schemas/Challenge_allOf'
      description: A challenge
      example:
        id: 507f1f77bcf86cd799439011
        name: awesome-challenge
        displayName: Awesome Challenge
        description: This challenge is awesome.
        websiteUrl: https://synapse.org/awesome-challenge
        status: active
        startDate: 2020-11-10
        endDate: 2020-12-31
        platformId: 507f1f77bcf86cd799439011
        fullName: dream/awesome-challenge
        ownerId: 507f1f77bcf86cd799439011
        createdAt: 2017-07-08T16:18:44-04:00
        updatedAt: 2017-07-08T16:18:44-04:00
      required:
      - createdAt
      - description
      - displayName
      - fullName
      - id
      - name
      - ownerId
      - status
      - updatedAt
      - websiteUrl
      title: Challenge
      type: object
    PageOfChallenges:
      allOf:
      - $ref: '#/components/schemas/ResponsePageMetadata'
      - $ref: '#/components/schemas/PageOfChallenges_allOf'
      description: A page of Challenges
      title: PageOfChallenges
      type: object
    Error:
      description: Problem details (tools.ietf.org/html/rfc7807)
      properties:
        title:
          description: A human readable documentation for the problem type
          title: title
          type: string
        status:
          description: The HTTP status code
          title: status
          type: integer
        detail:
          description: A human readable explanation specific to this occurrence of
            the problem
          title: detail
          type: string
        type:
          description: An absolute URI that identifies the problem type
          title: type
          type: string
      required:
      - status
      - title
      title: Error
      type: object
    EmptyObject:
      description: Empty JSON object
      type: object
...
    OrganizationCreateResponse:
      description: The unique identifier of the org created
      example:
        id: 507f1f77bcf86cd799439011
      properties:
        id:
          description: The unique identifier of an account
          example: 507f1f77bcf86cd799439011
          title: AccountId
          type: string
      required:
      - id
      title: OrganizationCreateResponse
      type: object
    OrganizationCreateRequest:
      description: The information required to create an org account
      example:
        login: awesome-org
      properties:
        login:
          type: string
      required:
      - login
      type: object
...
    Organization:
      allOf:
      - $ref: '#/components/schemas/OrganizationCreateResponse'
      - $ref: '#/components/schemas/OrganizationCreateRequest'
      - $ref: '#/components/schemas/Organization_allOf'
      description: An organization
      example:
        id: 507f1f77bcf86cd799439011
        login: awesome-org
        name: Awesome User
        email: awesome.user@example.com
        avatarUrl: https://example.com/awesome-avatar.png
        description: A great organization
        createdAt: 2017-07-08T16:18:44-04:00
        updatedAt: 2017-07-08T16:18:44-04:00
        type: Organization
      required:
      - avatarUrl
      - id
      - login
      - type
      title: Organization
      type: object
    PageOfOrganizations:
      allOf:
      - $ref: '#/components/schemas/ResponsePageMetadata'
      - $ref: '#/components/schemas/PageOfOrganizations_allOf'
      description: A page of Organizations
      title: PageOfOrganizations
      type: object
    AccountLogin:
      description: The user or organization account name
      example: awesome-user
      maxLength: 25
      minLength: 3
      pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
      type: string
...
    ResponsePageMetadata_paging:
      description: Links to navigate to different pages of results
      properties:
        next:
          description: Link to the next page of results
          format: uri
          title: next
          type: string
      title: ResponsePageMetadata_paging
      type: object
    Challenge_allOf:
      properties:
        fullName:
          type: string
        ownerId:
          description: The unique identifier of an account
          example: 507f1f77bcf86cd799439011
          title: AccountId
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      type: object
    PageOfChallenges_allOf:
      properties:
        challenges:
          description: An array of Challenges
          items:
            $ref: '#/components/schemas/Challenge'
          type: array
      required:
      - challenges
      type: object
...
    Organization_allOf:
      properties:
        name:
          nullable: true
          type: string
        email:
          description: An email address
          example: john.smith@example.com
          format: email
          type: string
        avatarUrl:
          example: https://example.com/awesome-avatar.png
          format: uri
          type: string
        description:
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        type:
          example: Organization
          type: string
      type: object
    PageOfOrganizations_allOf:
      properties:
        organizations:
          description: An array of Organizations
          items:
            $ref: '#/components/schemas/Organization'
          type: array
      required:
      - organizations
      type: object
...